requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2.13
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
importlib-metadata==4.11.4
Markdown==3.3.7
numpy==1.22.4
opencv-python==4.5.5.64
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.1.0
zipp==3.8.0
